// src/state/types.ts
export type SerializedBigNumber = string
// ...
export interface VaultApy {
  dailyApy: number
  weeklyApy: number
  monthlyApy: number
  yearlyApy: number
  yearlyApr: number
}